Add a simple GitHub action workflow (#4)
This simple GitHub action workflow is intended to test if it is possible to run the test environment on Github Actions. As it looks like, the test environment runs without any changes in Github Actions. It also can run with privileged mode and without direct_io=false which means the docker-db can use all its capabilities.