Add doctor command and module
The doctor module and command provide a way, to check the status, of the
environment for running and using the library and commandline tools
of the exasol_integration_test_docker_environment package.

By providing the "health" analysis functionality as lib and command-line
tool, the user can choose when and how to check for potentional issues.
This allows the user to choose when to do the checks and allows
for time optimization, in time cirtical paths.

So ideally the doctor cli tool is run once to check the environment
before code is executed. Successive executions then can make
more reliable assumptions regarding their environment, without
the need, to check them all the time by themselfs.

If a more custom/flexible way of doing the checks is needed,
the users can make calls to the provided checks and analysis
functionalities using the doctor module from within the code.