Improve error message (#180)
Add health checks for the different stages of the framework

In order to not slow down execution time permanently, we have taken the decision
to factor out the health/status checks, so they can be run whenever it does make
sense e.g. once before execution is started, instead of having them as permanent
path of execution within the rest of the code base.

While discussing the execution flow etc. we realized that the checks itself
also need to be split up, so we can provide specific checks for each "execution phase".

Execution Phases:

* Stage0 (shell scripts)
 [only bash and standard posix tools, can be assumed to be available]

  - The health check for this phase is taken care of by the shell script (health.sh)

Stage1 (python code running within a docker container)

  - The health check for this phase is taken care of by the python based cli tool
     and the doctor.py package. 

Co-authored-by:  tomuben and tkilias