Fix CVE-2021-32559
The CVE-2021-32559 is caused by the transitive dependency `pywin32`,
which is actually not needed by this package.

`pywin32` is pulled in by the `docker` dependency, but is only needed on
windows platforms (which ain't supported by this package).

By tigthening the dependency constrains of the `docker` dependency, we can
get rid of the optional (not needed) transitive `pywin32` dependency.